backend/requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0
ibm-watson==6.1.0
python-dotenv==1.0.0
flask-cors==4.0.0
